requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.20.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
auto-py-to-exe==2.43.3
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.2.3
blinker==1.8.1
bottle==0.12.25
bottle-websocket==0.2.9
branca==0.7.2
brython==3.11.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cloudscraper==1.2.71
cmake==3.29.3
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
cytoolz==0.12.3
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.6
dnspython==2.6.1
Eel==0.16.0
email_validator==2.1.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.10.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gTTS==2.5.2
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-socks==0.9.1
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
loguru==0.7.2
lru-dict==1.2.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.4.2
nexmo==2.5.2
nltk==3.8.1
numpy==1.26.4
ollama==0.3.3
openai==1.27.0
opencage==2.4.0
opencv-python==4.9.0.80
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
parsel==1.9.1
parsimonious==0.10.0
pefile==2023.2.7
phonenumbers==8.13.44
pillow==10.3.0
pluggy==1.5.0
Protego==0.3.1
protobuf==5.27.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycoin==0.92.20220529
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
pygame==2.5.1
Pygments==2.18.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
pyTelegramBotAPI==4.22.1
pytest==8.3.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.1
python-multipart==0.0.9
python-socks==2.5.3
pytz==2024.1
pyunormalize==15.1.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
queuelib==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.11.2
selenium==4.23.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.29
sqlparse==0.5.0
starlette==0.37.2
sympy==1.13.1
telebot==0.0.5
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.20.1
tomli==2.0.1
ton-client==0.5.1
tonpy==0.0.0.1.2b0
tonsdk==1.0.15
toolz==0.12.1
torch==2.5.1
tqdm==4.66.4
transformers==4.46.1
trio==0.26.0
trio-websocket==0.11.1
twilio==9.3.2
Twisted==24.7.0
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
w3lib==2.2.1
watchfiles==0.21.0
web3==6.19.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xyzservices==2024.9.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
